resolves RENATUS-1016, resolves RENATUS-981

Closes RENATUS-1016 and RENATUS-981

See merge request 75f/android/renatus!585

+++ logic/src/main/java/a75f/io/logic/bo/building/system/dab/DabSystemController.java
@@ -635,12 +635,12 @@ public class DabSystemController extends SystemController
             //Primary and secondary have the same base damper opening now.
             HashMap damper = hayStack.read("point and damper and base and cmd and equipRef == \""+m.get("id").toString()+"\"");
             double damperPos = hayStack.readHisValById(damper.get("id").toString());
-            int normalizedDamperPos = (int) (damperPos + damperPos * targetPercent/100);
+            double normalizedDamperPos =  (damperPos + damperPos * targetPercent/100);
             HashMap normalizedPrimaryDamper = hayStack.read("point and damper and normalized and primary and cmd and equipRef == \""+m.get("id").toString()+"\"");
             HashMap normalizedSecondaryamper = hayStack.read("point and damper and normalized and secondary and cmd and equipRef == \""+m.get("id").toString()+"\"");
             CcuLog.d(L.TAG_CCU_SYSTEM,"normalizeAirflow"+" Equip: "+m.get("dis")+" ,damperPos :"+damperPos+" targetPercent: "+targetPercent+" normalizedDamperPos: "+normalizedDamperPos);
-            hayStack.writeHisValById(normalizedPrimaryDamper.get("id").toString(), (double)normalizedDamperPos);
-            hayStack.writeHisValById(normalizedSecondaryamper.get("id").toString(), (double)normalizedDamperPos);
+            hayStack.writeHisValById(normalizedPrimaryDamper.get("id").toString(), normalizedDamperPos);
+            hayStack.writeHisValById(normalizedSecondaryamper.get("id").toString(), normalizedDamperPos);
         }
